basic_auth v2.1.0 BasicAuth

Plug for adding basic authentication. Usage:

plug BasicAuth, use_config: {:your_app, :your_key}

Where :your_app and :your_key should refer to values in your application config.

In your configuration you can set values directly, eg

config :your_app, your_config: [
  username: "admin",
  password: "simple_password",
  realm: "Admin Area"
]

or choose to get one (or all) from environment variables, eg

config :basic_auth, my_auth_with_system: [
  username: {:system, "BASIC_AUTH_USERNAME"},
  password: {:system, "BASIC_AUTH_PASSWORD"},
  realm:    {:system, "BASIC_AUTH_REALM"}
]

Alternatively you can provide a custom function to the plug to authenticate the user any way you want, such as finding the user from a database.

plug BasicAuth, callback: &User.find_by_username_and_password/3

Where :callback is your custom authentication function that takes a conn, username and a password and returns a conn. Your function must return Plug.Conn.halt(conn) if authentication fails, otherwise you can use Plug.Conn.assign(conn, :current_user, ...) to enhance the conn with variables or session for your controller.

Functions

call(conn, options)

init(arg1)